드론 격납고 개소 등록, 시설물 administ_zone값 입력 추가

+++ danusys-commons-api/src/main/java/com/danusys/web/commons/api/model/Station.java
@@ -28,7 +28,7 @@ public class Station {
     private String stationName;
 
     @Column(nullable = false)
-    private int stationKind;
+    private Long stationKind;
 
     @Column
     private String administZone;
